cli: retire the import-pdf wizard, redirect to the Config Builder TUI
The 13-step import-pdf -wizard is superseded by the full terminal Config
Builder (gophertrunk config / config tui), which edits every section and can
import PDF/CSV itself. So:
- Delete config_wizard.go/config_template.go (+ tests).
- import-pdf -wizard now launches the TUI; bare 'config' already does.
- Relocate the wizard's path helpers (~ / $VAR / %VAR%, default-path) into
  internal/configbuilder so the TUI's Save-As and -config flag accept the
  same operator-typed paths.
- Fix the config-load hint to point at 'gophertrunk config'.
